move array making function to the useFxApi hook
